requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.16
django-admin-list-filter-dropdown==1.0.3
django-ajax-datatable==4.4.4
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-bootstrap-datepicker-plus==5.0.3
django-comments-dab==2.8.0
django-cors-headers==3.14.0
django-countries==7.5.1
django-currentuser==0.5.3
django-import-export==3.0.2
django-jazzmin==2.6.0
django-model-utils==4.3.1
django-modeladmin-reorder==0.3.1
django-phonenumber-field==7.0.0
django-scheduler==0.10.1
django-select2==8.0.0
django-smart-selects==1.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
icalendar==5.0.4
jsonfield==3.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.0
phonenumbers==8.12.54
Pillow >=9
pydantic==1.10.6
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
six==1.16.0
sorl-thumbnail==12.9.0
sqlparse==0.4.2
swapper==1.3.0
tablib==3.3.0
typing_extensions==4.5.0
tzdata==2022.7
xlrd==2.0.1
xlwt==1.3.0
